fix:(service and migration problem) change sale_items table's column type from integer('quantity') to decimal('quantity', 10, 2)

+++ app/Services/SaleService.php
@@ -28,7 +28,7 @@ class SaleService
             foreach ($items as $item) {
 
                 $productId = $item['product_id'];
-                $quantity = (int) $item['quantity'];
+                $quantity = (float) $item['quantity'];
 
                 $product = Product::lockForUpdate()->find($productId);
 
@@ -72,7 +72,7 @@ class SaleService
             foreach ($items as $item) {
 
                 $product = $productsMap[$item['product_id']];
-                $quantity = (int) $item['quantity'];
+                $quantity = (float) $item['quantity'];
 
                 $price = (int) $product->price;
                 $subtotal = $price * $quantity;
